+++ src/mqtt_client.rs
@@ -2,14 +2,21 @@ use crate::{
     de::{deserialize::ReceivedPacket, PacketReader},
     ser::serialize,
     session_state::SessionState,
-    Property, {debug, error, info},
+    Property, {debug, error, info, warn},
 };
 
 use embedded_nal::{nb, IpAddr, SocketAddr, TcpClientStack};
+use embedded_time::{
+    duration::{Extensions, Seconds},
+    Clock,
+};
 
 use heapless::String;
 
 use core::str::FromStr;
+
+/// The default duration to wait for a ping response from the broker.
+const PING_TIMEOUT: embedded_time::duration::Seconds = embedded_time::duration::Seconds(5);
 
 /// The quality-of-service for an MQTT message.
 #[derive(Debug, Copy, Clone, PartialEq)]
@@ -25,7 +32,7 @@ pub enum QoS {
 }
 
 /// Possible errors encountered during an MQTT connection.
-#[derive(Debug, Copy, Clone, PartialEq)]
+#[derive(Debug, PartialEq)]
 pub enum Error<E> {
     Network(E),
     WriteFail,
@@ -35,6 +42,7 @@ pub enum Error<E> {
     Failed(u8),
     Protocol(ProtocolError),
     SessionReset,
+    Clock(embedded_time::clock::Error),
 }
 
 /// Errors that are specific to the MQTT protocol implementation.
@@ -53,9 +61,15 @@ pub enum ProtocolError {
     InvalidProperty,
 }
 
-impl<E> From<E> for Error<E> {
-    fn from(e: E) -> Error<E> {
-        Error::Network(e)
+impl<E> From<embedded_time::clock::Error> for Error<E> {
+    fn from(clock: embedded_time::clock::Error) -> Self {
+        Error::Clock(clock)
+    }
+}
+
+impl<E> From<ProtocolError> for Error<E> {
+    fn from(error: ProtocolError) -> Self {
+        Error::Protocol(error)
     }
 }
 
@@ -81,29 +95,35 @@ mod sm {
 use sm::{Context, Events, StateMachine, States};
 
 /// The general structure for managing MQTT via Minimq.
-pub struct Minimq<N, const T: usize>
+pub struct Minimq<N, C, const T: usize>
 where
     N: TcpClientStack,
+    C: Clock,
 {
-    pub client: MqttClient<N, T>,
+    pub client: MqttClient<N, C, T>,
     packet_reader: PacketReader<T>,
 }
 
 /// A client for interacting with an MQTT Broker.
-pub struct MqttClient<N: TcpClientStack, const T: usize> {
+pub struct MqttClient<N: TcpClientStack, C: Clock, const T: usize> {
     network_stack: N,
+    clock: C,
     socket: Option<N::TcpSocket>,
-    session_state: SessionState,
+    session_state: SessionState<C>,
     connection_state: StateMachine<Context>,
 }
 
-impl<N, const T: usize> MqttClient<N, T>
+impl<N, C, const T: usize> MqttClient<N, C, T>
 where
     N: TcpClientStack,
+    C: Clock,
 {
     fn process(&mut self) -> Result<(), Error<N::Error>> {
         // Potentially update the state machine depending on the current socket connection status.
-        if !self.socket_is_connected()? {
+        if !self
+            .socket_is_connected()
+            .map_err(|err| Error::Network(err))?
+        {
             self.connection_state.process_event(Events::Disconnect).ok();
         } else {
             self.connection_state.process_event(Events::Connect).ok();
@@ -112,12 +132,18 @@ where
         match self.connection_state.state() {
             // In the RESTART state, we need to reopen the TCP socket.
             &States::Restart => {
-                if self.socket.is_some() {
-                    self.network_stack.close(self.socket.take().unwrap())?;
+                if let Some(socket) = self.socket.take() {
+                    self.network_stack
+                        .close(socket)
+                        .map_err(|err| Error::Network(err))?;
                 }
 
                 // Allocate a new socket to use and begin connecting it.
-                self.socket.replace(self.network_stack.socket()?);
+                self.socket.replace(
+                    self.network_stack
+                        .socket()
+                        .map_err(|err| Error::Network(err))?,
+                );
                 self.connection_state
                     .process_event(Events::GotSocket)
                     .unwrap();
@@ -151,12 +177,14 @@ where
                 let packet = serialize::connect_message(
                     &mut buffer,
                     self.session_state.client_id.as_str().as_bytes(),
-                    self.session_state.keep_alive_interval,
+                    self.session_state
+                        .keep_alive_interval
+                        .unwrap_or(0.seconds())
+                        .0 as u16,
                     &properties,
                     // Only perform a clean start if we do not have any session state.
                     !self.session_state.is_present(),
-                )
-                .map_err(|e| Error::Protocol(e))?;
+                )?;
 
                 info!("Sending CONNECT");
                 self.write(packet)?;
@@ -168,6 +196,8 @@ where
 
             _ => {}
         }
+
+        self.handle_timers()?;
 
         Ok(())
     }
@@ -195,9 +225,42 @@ where
                 if written != buf.len() {
                     Err(Error::WriteFail)
                 } else {
+                    self.session_state
+                        .register_transmission(self.clock.try_now()?);
                     Ok(())
                 }
             })
+    }
+
+    /// Configure the MQTT keep-alive interval.
+    ///
+    /// # Note
+    /// This must be completed before connecting to a broker.
+    ///
+    /// # Note
+    /// The broker may override the requested keep-alive interval. Any value requested by the
+    /// broker will be used instead.
+    ///
+    /// # Args
+    /// * `interval` - The keep-alive interval in seconds. A ping will be transmitted if no other
+    /// messages are sent within 50% of the keep-alive interval.
+    pub fn set_keepalive_interval(
+        &mut self,
+        interval: impl Into<Seconds<u32>>,
+    ) -> Result<(), Error<N::Error>> {
+        let interval = interval.into();
+        if interval.0 > u16::MAX as u32 {
+            return Err(ProtocolError::Invalid.into());
+        }
+        match self.connection_state.state() {
+            &States::Active => Err(Error::NotReady),
+            _ => {
+                self.session_state
+                    .keep_alive_interval
+                    .replace(interval.into());
+                Ok(())
+            }
+        }
     }
 
     /// Subscribe to a topic.
@@ -221,8 +284,7 @@ where
         let packet_id = self.session_state.get_packet_identifier();
 
         let mut buffer: [u8; T] = [0; T];
-        let packet = serialize::subscribe_message(&mut buffer, topic, packet_id, properties)
-            .map_err(|e| Error::Protocol(e))?;
+        let packet = serialize::subscribe_message(&mut buffer, topic, packet_id, properties)?;
 
         self.write(packet).and_then(|_| {
             info!("Subscribing to `{}`: {}", topic, packet_id);
@@ -248,7 +310,10 @@ where
     /// # Returns
     /// True if the client is connected to the broker.
     pub fn is_connected(&mut self) -> Result<bool, Error<N::Error>> {
-        Ok(self.socket_is_connected()? && self.session_state.connected)
+        Ok(self
+            .socket_is_connected()
+            .map_err(|err| Error::Network(err))?
+            && self.session_state.connected)
     }
 
     /// Publish a message over MQTT.
@@ -286,8 +351,7 @@ where
         );
 
         let mut buffer: [u8; T] = [0; T];
-        let packet = serialize::publish_message(&mut buffer, topic, data, properties)
-            .map_err(|e| Error::Protocol(e))?;
+        let packet = serialize::publish_message(&mut buffer, topic, data, properties)?;
         self.write(packet)
     }
 
@@ -310,7 +374,7 @@ where
         f: &mut F,
     ) -> Result<(), Error<N::Error>>
     where
-        F: FnMut(&mut MqttClient<N, T>, &'a str, &[u8], &[Property<'a>]),
+        F: FnMut(&mut MqttClient<N, C, T>, &'a str, &[u8], &[Property<'a>]),
     {
         if !self.session_state.connected {
             if let ReceivedPacket::ConnAck(acknowledge) = packet {
@@ -344,7 +408,9 @@ where
                                 String::from_str(id).or(Err(Error::ProvidedClientIdTooLong))?;
                         }
                         Property::ServerKeepAlive(keep_alive) => {
-                            self.session_state.keep_alive_interval = keep_alive;
+                            self.session_state
+                                .keep_alive_interval
+                                .replace((keep_alive as u32).seconds());
                         }
                         _prop => info!("Ignoring property: {:?}", _prop),
                     };
@@ -391,12 +457,55 @@ where
                 Ok(())
             }
 
+            ReceivedPacket::PingResp => {
+                // Cancel the ping response timeout.
+                if self.session_state.ping_timeout.is_some() {
+                    self.session_state.ping_timeout = None;
+                } else {
+                    warn!("Got unexpected ping response");
+                }
+
+                Ok(())
+            }
+
             _ => Err(Error::Unsupported),
         }
     }
+
+    fn handle_timers(&mut self) -> Result<(), Error<N::Error>> {
+        // If we are not connected, there's no session state to manage.
+        if self.connection_state.state() != &States::Active {
+            return Ok(());
+        }
+
+        let now = self.clock.try_now()?;
+
+        if let Some(timeout) = self.session_state.ping_timeout {
+            if now > timeout {
+                // Reset network connection.
+                self.connection_state.process_event(Events::Disconnect).ok();
+            }
+        } else {
+            // Check if we need to ping the server.
+            if self.session_state.ping_is_due(&now) {
+                let mut buffer: [u8; T] = [0; T];
+
+                // Note: The ping timeout is set at this point so that it's running even if we fail
+                // to write the ping message. This is intentional incase the underlying transport
+                // mechanism has stalled. The ping timeout will then allow us to recover the
+                // underlying TCP connection.
+                self.session_state.ping_timeout.replace(now + PING_TIMEOUT);
+
+                let packet = serialize::ping_req_message(&mut buffer)?;
+                self.write(packet)?;
+            }
+        }
+
+        Ok(())
+    }
 }
 
-impl<N: TcpClientStack, const T: usize> Minimq<N, T> {
+impl<N: TcpClientStack, C: Clock, const T: usize> Minimq<N, C, T> {
     /// Construct a new MQTT interface.
     ///
     /// # Args
@@ -404,11 +513,17 @@ impl<N: TcpClientStack, const T: usize> Minimq<N, T> {
     /// * `client_id` The client ID to use for communicating with the broker. If empty, rely on the
     ///   broker to automatically assign a client ID.
     /// * `network_stack` - The network stack to use for communication.
+    /// * `clock` - The clock to use for managing MQTT state timing.
     ///
     /// # Returns
     /// A `Minimq` object that can be used for publishing messages, subscribing to topics, and
     /// managing the MQTT state.
-    pub fn new(broker: IpAddr, client_id: &str, network_stack: N) -> Result<Self, Error<N::Error>> {
+    pub fn new(
+        broker: IpAddr,
+        client_id: &str,
+        network_stack: N,
+        clock: C,
+    ) -> Result<Self, Error<N::Error>> {
         let session_state = SessionState::new(
             broker,
             String::from_str(client_id).or(Err(Error::ProvidedClientIdTooLong))?,
@@ -417,6 +532,7 @@ impl<N: TcpClientStack, const T: usize> Minimq<N, T> {
         let minimq = Minimq {
             client: MqttClient {
                 network_stack,
+                clock,
                 socket: None,
                 session_state,
                 connection_state: StateMachine::new(Context),
@@ -434,7 +550,7 @@ impl<N: TcpClientStack, const T: usize> Minimq<N, T> {
     /// topic, message, and list of proprties (in that order).
     pub fn poll<F>(&mut self, mut f: F) -> Result<(), Error<N::Error>>
     where
-        for<'a> F: FnMut(&mut MqttClient<N, T>, &'a str, &[u8], &[Property<'a>]),
+        for<'a> F: FnMut(&mut MqttClient<N, C, T>, &'a str, &[u8], &[Property<'a>]),
     {
         self.client.process()?;
 
@@ -468,16 +584,13 @@ impl<N: TcpClientStack, const T: usize> Minimq<N, T> {
 
             // Handle any received packets.
             while self.packet_reader.packet_available() {
-                let packet = ReceivedPacket::parse_message(&self.packet_reader)
-                    .map_err(|e| Error::Protocol(e))?;
+                let packet = ReceivedPacket::parse_message(&self.packet_reader)?;
 
                 info!("Received {:?}", packet);
 
                 let result = self.client.handle_packet(packet, &mut f);
 
-                self.packet_reader
-                    .pop_packet()
-                    .map_err(|e| Error::Protocol(e))?;
+                self.packet_reader.pop_packet()?;
 
                 // If there was an error, return it now. Note that we ensure the packet is removed
                 // from buffering after processing even in error conditions..