+++ src/ser/serialize.rs
@@ -62,6 +62,10 @@ pub fn connect_message<'a, 'b>(
     packet.finalize(MessageType::Connect, 0)
 }
 
+pub fn ping_req_message<'a>(dest: &'a mut [u8]) -> Result<&'a [u8], Error> {
+    ReversedPacketWriter::new(dest).finalize(MessageType::PingReq, 0x00)
+}
+
 pub fn publish_message<'a, 'b, 'c>(
     dest: &'b mut [u8],
     topic: &'a str,
@@ -204,4 +208,15 @@ fn serialize_connect() {
     let message = connect_message(&mut buffer, client_id, 10, &[], true).unwrap();
 
     assert_eq!(message, good_serialized_connect)
+}
+
+#[test]
+fn serialize_ping_req() {
+    let good_ping_req: [u8; 2] = [
+        0xc0, // Ping
+        0x00, // Remaining length (0)
+    ];
+
+    let mut buffer: [u8; 1024] = [0; 1024];
+    assert_eq!(ping_req_message(&mut buffer).unwrap(), good_ping_req);
 }